pyproject.toml
[tool.mypy]
mypy_path = "tense"
python_version = "3.10"
exclude = "__init__.py"

strict = true
warn_unused_configs = true
disallow_subclassing_any = false

show_error_codes = true
show_column_numbers = true

follow_imports = "skip"
ignore_missing_imports = true
no_site_packages = true

[tool.isort]
profile = "black"
py_version = 310
src_paths = ["tense", "tests", "examples"]

[tool.black]
target-version = ["py310"]
include = '\.pyi?$'

[tool.poetry]
name = "tense"
version = "1.0.0b"
description = "Time Processing Tool"
authors = ["Animatea <animatea.programming@gmail.com>"]
readme = "PYPI_README.md"
license = "Apache-2.0"
homepage = "https://github.com/Animatea/tense"
repository = "https://github.com/Animatea/tense"
documentation = "https://animatea.github.io/tense/"
include = [
    "LICENSE", "assets", "i18n",
]
keywords = ["parser", "tool", "tense", "time", "utility"]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Development Status :: 3 - Alpha",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: Implementation :: CPython",
    "License :: OSI Approved :: Apache Software License",
    "Typing :: Typed",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]
packages = [
    { include = "tense" },
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Animatea/tense/issues"
"Documentation" = "https://animatea.github.io/tense/"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
parse = "1.19.0"
PyHamcrest = "2.0.3"
pytest = "7.1.2"
behave = "1.2.6"
black = "22.3.0"
isort = "5.10.1"
mypy = "0.961"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.core.masonry.api"
